+++ validator/src/main.rs
@@ -1086,7 +1086,7 @@ pub fn main() {
 
     let tpu_connection_pool_size = value_t_or_exit!(matches, "tpu_connection_pool_size", usize);
     let tpu_max_connections_per_ipaddr_per_minute =
-        value_t_or_exit!(matches, "tpu_max_connections_per_ipaddr_per_minute", u32);
+        value_t_or_exit!(matches, "tpu_max_connections_per_ipaddr_per_minute", u64);
 
     let shrink_ratio = value_t_or_exit!(matches, "accounts_shrink_ratio", f64);
     if !(0.0..=1.0).contains(&shrink_ratio) {
